requirements.txt
requests
pytz
pyyaml
prompt_toolkit
llama-index
llama-index-llms-ollama 
python-dotenv
